Restructure to support Python 3.7
