+++ samples/test_azure_event_hubs_mgmt_op_async.py
@@ -11,7 +11,6 @@ import pytest
 import sys
 
 import uamqp
-from uamqp import async as a_uamqp
 from uamqp import authentication
