+++ samples/test_azure_iothub_receive_async.py
@@ -20,7 +20,6 @@ except Exception:
     from urllib.parse import quote, quote_plus, urlencode
 
 import uamqp
-from uamqp import async as a_uamqp
 from uamqp import authentication, errors, address
 
 
@@ -62,7 +61,7 @@ def _build_iothub_amqp_endpoint_from_target(target):
 
 
 async def _receive_mesages(conn, source, auth):
-    receive_client = a_uamqp.ReceiveClientAsync(source, auth=auth, debug=True, timeout=1000, prefetch=1)
+    receive_client = uamqp.ReceiveClientAsync(source, auth=auth, debug=True, timeout=1000, prefetch=1)
     try:
         await receive_client.open_async(connection=conn)
         batch = await receive_client.receive_message_batch_async(max_batch_size=1)
@@ -83,7 +82,7 @@ async def test_iothub_client_receive_async(live_iothub_config):
     source = 'amqps://' + live_iothub_config['hostname'] + operation
     log.info("Source: {}".format(source))
 
-    async with a_uamqp.ConnectionAsync(live_iothub_config['hostname'], auth, debug=True) as conn:
+    async with uamqp.ConnectionAsync(live_iothub_config['hostname'], auth, debug=True) as conn:
         tasks = [
             _receive_mesages(conn, source + '0', auth),
             _receive_mesages(conn, source + '1', auth)
