+++ uamqp/_async/__init__.py
@@ -6,8 +6,6 @@
 
 from .connection_async import ConnectionAsync
 from .session_async import SessionAsync
-from .client_async import AMQPClientAsync, SendClientAsync, ReceiveClientAsync
 from .sender_async import MessageSenderAsync
 from .receiver_async import MessageReceiverAsync
 from .mgmt_operation_async import MgmtOperationAsync
-from .authentication_async import SASTokenAsync
