+++ uamqp/authentication/__init__.py
@@ -1,0 +1,12 @@
+#-------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+#--------------------------------------------------------------------------
+
+from .common import AMQPAuth, SASLPlain, SASLAnonymous
+from .cbs_auth import TokenRetryPolicy, CBSAuthMixin, SASTokenAuth
+try:
+    from .cbs_auth_async import CBSAsyncAuthMixin, SASTokenAsync
+except (ImportError, SyntaxError):
+    pass  # No Python 3.4+ support
